refactor: move e2e test inside to tests
